backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=1.60.0
supabase>=2.11.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
PyJWT>=2.8.0
